+++ othellia/minimax.py
@@ -3,8 +3,8 @@ from typing import Callable, Type
 
 import numpy as np
 
+from othellia.game import Game
 from settings.values import BLACK_VALUE
-from src.game import Game
 
 
 def minimax(
