+++ othellia/sprites.py
