--- tests/__init__.py
@@ -1,3 +1,0 @@
-import sys
-
-sys.path.append("./src")
